pyproject.toml
[tool.poetry]
name = "twitter"
version = "0.1.0"
description = "A twiiter api clone"
authors = ["hephhay <faradaydanfard@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/hephhay/twitter/"
repository = "https://github.com/hephhay/twitter/"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.2"
psycopg2 = "^2.9.4"
flake8 = "^5.0.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
djoser = "^2.1.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
celery = {extras = ["auth", "librabbitmq", "msgpack", "redis"], version = "^5.2.7"}
django-celery-results = "^2.4.0"
redis = "^4.4.2"
hiredis = "^2.1.1"
channels-redis = "^4.0.0"
whitenoise = "^6.3.0"
django-cors-headers = "^3.13.0"
pillow = "^9.4.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
isort = {extras = ["pipfile-deprecated-finder", "requirements-deprecated-finder"], version = "^5.10.1"}
flake8 = "^5.0.4"
pytest = "^7.1.3"
doc8 = "^1.0.0"
types-docutils = "^0.19.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]

